A3

+++ pe_exchange.c
@@ -188,12 +188,12 @@ int main(int argc, char **argv)
 						send_invalid_message_to_current_trader(t, invalid_message);
 						continue;
 					}
-					long new_qty = atoi(qty);
+					long new_qty = atol(qty);
 					char *price = strtok(NULL, ";");
 					if (price == NULL) {
 						send_invalid_message_to_current_trader(t, invalid_message);
 					}
-					long new_price = atoi(price);
+					long new_price = atol(price);
 					int updated = update_order(book, order_id, new_qty, new_price, t);
 					if (updated) {
 						char* msg = malloc(sizeof(char) * INPUT_LENGTH);
